requirements.txt
streamlit
pandas
pydeck
openelectricity
paho-mqtt
